Some fixes

+++ src/main/java/school/faang/user_service/service/goal/GoalService.java
@@ -1,20 +1,21 @@
 package school.faang.user_service.service.goal;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.repository.goal.GoalRepository;
 
 
 @Component
+@RequiredArgsConstructor
 public class GoalService {
     private final GoalRepository goalRepository;
 
-    @Autowired
-    public GoalService(GoalRepository goalRepository) {
-        this.goalRepository = goalRepository;
-    }
 
     public void deleteGoal(long goalId) {
+        if(!goalRepository.existsById(goalId)){
+            throw new DataValidationException("Такой цели нет");
+        }
         goalRepository.deleteById(goalId);
     }
 }
